requirements.txt
aiofiles
fastapi
uvicorn
typer
sqlalchemy
databases[sqlite]
gym
pyserial
python-multipart
docker